refactor: The directory containing generic class files is just called `classes` now
