+++ simple_games/classes/player.py
