+++ simple_games/classes/strategy_interface.py
