+++ simple_games/classes/tournament.py
