+++ simple_games/graph_coloring/classes/gc_data_strategy.py
@@ -12,7 +12,7 @@ from random import choice
 
 import sys
 sys.path.append("../..")
-from src.strategy_interface import StrategyInterface
+from classes.strategy_interface import StrategyInterface
 
 class GCDataStrategy(StrategyInterface):
