+++ simple_games/toothpick_takeaway/classes/toothpick_ruleset.py
@@ -10,7 +10,7 @@
 
 import sys
 sys.path.append("../..")
-from src.ruleset_interface import RulesetInterface
+from classes.ruleset_interface import RulesetInterface
 
 class ToothpickRuleset(RulesetInterface):
