+++ simple_games/toothpick_takeaway/classes/toothpick_strategy.py
@@ -15,7 +15,7 @@ import random
 
 import sys
 sys.path.append("../..")
-from src.strategy_interface import StrategyInterface
+from classes.strategy_interface import StrategyInterface
 
 class ToothpickStrategy(StrategyInterface):
