+++ simple_games/unbalanced_rook/classes/rook_strategy.py
@@ -14,7 +14,7 @@
 import random
 import sys
 sys.path.append("../..")
-from src.strategy_interface import StrategyInterface
+from classes.strategy_interface import StrategyInterface
 
 class RookStrategy(StrategyInterface):
